Renamed XmlService to CloudStorageService, added path to bucket in application.yml file
