+++ Backend/microservices/data/pom.xml
@@ -57,6 +57,7 @@
         </dependency>
 
 
+
     </dependencies>
 
 </project>
